Implement save/load for ROLVector
As part of pyrol serialisation, an overridden Vector implementor needs
to support pickling. Since we can't (shouldn't?) simply pickle a
Firedrake vector, we use DumbCheckpoint and pickle just the filename
containing the actual data.

There is a complication with this approach, that DumbCheckpoint can't
be used to unpickle, because we don't know the mesh/function space on
which to load the data. Until firedrake#1917 is merged, we can get
around this by registering all unpickled vectors, then their data can
be loaded from disk once the mesh/function space are initialised.