Convert ROLObjective to handle update flag
This lets us revert to a previous iterate without re-evaluating.

If the last call to .update was a Revert, then we need to re-evaluate
the reduced functional anyway, to make sure the tape is in the right
place for the gradient evaluation. It might be possible that this
never occurs, but best be safe!