Save and restore the tape cache on ROLObjective calls
The addition of the ROL update flags should allow us to minimise
recomputation of the ReducedFunctional and its gradient. In the case
of a reverted objective update, we need to make sure the tape of the
ReducedFunctional is in sync with the control values we've loaded in,
because of the following possible call sequence if we don't
specifically handle this:

1. obj.update(x, Accept) # rf and obj are in sync for x
2. obj.update(y, Trial)  # rf and obj are in sync for y
3. obj.update(x, Revert) # rf still points to y, but obj is x
4. obj.gradient() # calculated from rf, gives gradient at y

We cache all the BlockVariable data on the tape at an Accept call, and
can recall that onto the tape when Revert is invoked, thus keeping the
ReducedFunctional and the objective in sync.